Deduplicate one record and actually write some output
While the warcdedupe binary now actually writes some output, it will only
process one record from the input and hasn't been tested very far.

warcio: add support for writing a record from memory, and require callers
to specify the compression for a record when reading. Also add a header
method that handles the different between URI syntax between WARC 1.0 and
1.1.

warcdedupe: introduce a Digester abstraction to support various hashing
methods, currently supporting the same BLAKE3 digest. Begin requiring the
input to be seekable in order to efficiently be able to copy records to the
output, and implement output both of revisit records and copying records
unmodified to the output.