Big changes to header field representation
We now use an enum that knows about standardized headers, which should be
more compact for those standardized values and opens the door to making the
code more aware of the semantics of standard fields. Judicious use of
conversion traits also improves the ergonomics.

The small downside is needing to do code generation in build.rs, but that's
relatively simple and on balance worthwhile since it enables the other new
features.