warcio: allow users to provide and retrieve input buffers
Allocating large buffers for reading into turns out to be costly because they
are zeroed on allocation. buf_redux's buffers are uninitialized on creation
so much less costly, but it's even better to support reusing a buffer between
records- do so by adding a function to a record to get the input and buffer
back out.