Optimize the Eq impl for FieldName slightly
Recognizing that we only need to do an actual string comparison when either
variant is an unrecognized one, in most cases this can directly compare the
enum discriminant.

This also makes the Ord impl more obviously correct by delegating to Ord for
UncasedStr rather than implying that PartialOrd is never None.