Refactor header field parsing to avoid regexes
A regex is a quick and easy way to match chunks of bytes, but the overhead
seems significant in this application where an individual match tends to be
no more than a few tens of bytes.

I tried to use `httparse` to implement header parsing because WARC headers
are specified to be compliant with HTTP, but httparse notably lacks support
for line folding because more recent HTTP specifications deprecate it. Instead
we need a custom parser.